+++ src/providers/bitbucket.ts
@@ -1,0 +1,15 @@
+import { defineProvider, parseGitURI } from "../_utils";
+
+export default defineProvider((input, options) => {
+  const parsed = parseGitURI(input);
+  return {
+    name: parsed.repo.replace("/", "-"),
+    version: parsed.ref,
+    subdir: parsed.subdir,
+    headers: {
+      authorization: options.auth ? `Bearer ${options.auth}` : undefined,
+    },
+    url: `https://bitbucket.com/${parsed.repo}/src/${parsed.ref}${parsed.subdir}`,
+    tar: `https://bitbucket.org/${parsed.repo}/get/${parsed.ref}.tar.gz`,
+  };
+});
