+++ src/providers/github.ts
@@ -1,0 +1,24 @@
+import { defineProvider, parseGitURI } from "../_utils";
+
+export default defineProvider((input, options) => {
+  const parsed = parseGitURI(input);
+
+  // https://docs.github.com/en/rest/repos/contents#download-a-repository-archive-tar
+  // TODO: Verify solution for github enterprise
+  const githubAPIURL = process.env.GIGET_GITHUB_URL || "https://api.github.com";
+
+  return {
+    name: parsed.repo.replace("/", "-"),
+    version: parsed.ref,
+    subdir: parsed.subdir,
+    headers: {
+      Authorization: options.auth ? `Bearer ${options.auth}` : undefined,
+      Accept: "application/vnd.github+json",
+      "X-GitHub-Api-Version": "2022-11-28",
+    },
+    url: `${githubAPIURL.replace("api.github.com", "github.com")}/${
+      parsed.repo
+    }/tree/${parsed.ref}${parsed.subdir}`,
+    tar: `${githubAPIURL}/repos/${parsed.repo}/tarball/${parsed.ref}`,
+  };
+});
