+++ src/providers/gitlab.ts
@@ -1,0 +1,18 @@
+import { defineProvider, parseGitURI } from "../_utils";
+
+export default defineProvider((input, options) => {
+  const parsed = parseGitURI(input);
+  const gitlab = process.env.GIGET_GITLAB_URL || "https://gitlab.com";
+  return {
+    name: parsed.repo.replace("/", "-"),
+    version: parsed.ref,
+    subdir: parsed.subdir,
+    headers: {
+      authorization: options.auth ? `Bearer ${options.auth}` : undefined,
+      // https://gitlab.com/gitlab-org/gitlab/-/commit/50c11f278d18fe1f3fb12eb595067216bb58ade2
+      "sec-fetch-mode": "same-origin",
+    },
+    url: `${gitlab}/${parsed.repo}/tree/${parsed.ref}${parsed.subdir}`,
+    tar: `${gitlab}/${parsed.repo}/-/archive/${parsed.ref}.tar.gz`,
+  };
+});
