+++ src/providers/sourcehut.ts
@@ -1,0 +1,15 @@
+import { defineProvider, parseGitURI } from "../_utils";
+
+export default defineProvider((input, options) => {
+  const parsed = parseGitURI(input);
+  return {
+    name: parsed.repo.replace("/", "-"),
+    version: parsed.ref,
+    subdir: parsed.subdir,
+    headers: {
+      authorization: options.auth ? `Bearer ${options.auth}` : undefined,
+    },
+    url: `https://git.sr.ht/~${parsed.repo}/tree/${parsed.ref}/item${parsed.subdir}`,
+    tar: `https://git.sr.ht/~${parsed.repo}/archive/${parsed.ref}.tar.gz`,
+  };
+});
